Move the mimic-QPU validation to RemoteBackend.run() (#888)
* Move the mimic-QPU validation to RemoteBackend.run()

* fix test cov

* apply comments

* fix flake

---------
